added the vaiaubles lib
