+++ analysis/dataset_definition_dx_to_med_time.py
@@ -12,23 +12,8 @@ from codelists import (
     methylphenidate_codelist
 )
 
-def last_matching_event(events, codelist, where=True):
-    """Select the last matching SNOMED CT event from specified codelist
+from variables_library import last_matching_event
 
-    Args:
-        events (e.g., clinical_events or ): Many rows per patient event frame to select last matching event from codelist from
-        codelist (codelist): Clinical codelist, must be using snomedct codes
-        where (bool, optional): _description_. Defaults to True.
-
-    Returns:
-        patient frame: One row per patient frame, with the last matching event from codelist
-    """
-    return (
-        events.where(where)
-        .where(events.snomedct_code.is_in(codelist))
-        .sort_by(events.date)
-        .last_for_patient()
-    )
 
 dataset = create_dataset()
 dataset.configure_dummy_data(population_size=5)
